Add files via upload
People get to know one another by sharing their
ideas, thoughts, and experiences with those in their immediate
surroundings. There are numerous methods for accomplishing
this, the most effective of which is the gift of ”Speech.” Speech
enables all people to communicate their ideas effectively and
to comprehend one another. It will be unfair if we fail to
take into consideration those who are denied this priceless gift:
the dumb and deaf. The preferred method of communication
in these situations has continued to be human hand contact.
Things that have been first challenging or unattainable for people
with disabilities are now regularly available to them and can
be accessed by them with ease. Artificial intelligence made it
possible for people with disabilities to live in a society where
their challenges are acknowledged and taken into account (AI).
Technological advancements have made it possible for technology
to adjust and transform the world into a more open community.
There is a certain sense of being human as AI directly correlates
individuals, including people with and without impairments.
hearing voice in the preferred language so that a message can be
delivered it to normal people to build a model that is trained
on various hand motions we are using a convolution neural
network and deep learning on the basis of this model an app
is created with the help of this app person who is deaf or dumb
can communicate using postures that are translated into speech
and human-understandable words
Index Terms—CNN, Image Processing, dumb and deaf, artificial Intelligence,gesture recognition,convertor gor image to
text,model fiting